HashMap: Use explicit cast to get the hash code

+++ include/nstd/HashSet.h
@@ -75,7 +75,7 @@ public:
   Iterator find(const T& key)
   {
     if(!data) return _end;
-    size_t hashCode = key;
+    size_t hashCode = (size_t)key;
     Item* item = data[hashCode % capacity];
     while(item)
     {
@@ -118,7 +118,7 @@ public:
       }
     }
 
-    size_t hashCode = key;
+    size_t hashCode = (size_t)key;
     VERIFY(new(item) Item(key) == item);
     //item->Item::Item(key);
     Item** cell;
